refactoring cm api, offer creation
