+++ lib/cm/client.js
@@ -2,13 +2,15 @@
  * Created by kugajjar on 6/3/15.
  */
 
+'use strict';
+
 var crypto          = require('crypto'),
     request         = require('request'),
     HttpStatus      = require('http-status-codes'),
     url             = require('url'),
     Q               = require('q');
 
-module.exports = function api_client(nemo) {
+module.exports = function client(nemo) {
 
     var config = nemo._config.get('cm');
 
@@ -75,8 +77,6 @@ module.exports = function api_client(nemo) {
                     d.reject(err);
                 }
             }
-
-            d.resolve();
 
         });
