+++ lib/cm/cm.js
@@ -6,18 +6,19 @@
 
 var affiliate_offer         = require('../../providers/affiliate_offer.json').affiliate_offer,
     Q                       = require('q'),
-    client                  = require('./api_client');
+    utils                   = require('../utils/utils'),
+    client                  = require('./client');
 
 require('date-utils');
 
-module.exports = function offer(nemo) {
+module.exports = function cm(nemo) {
 
     var setup_affiliate_offer = function () {
 
         var deferred = Q.defer();
 
-        affiliate_offer.bml_id = generate_random_number(10);
-        affiliate_offer.shopping_title = '$$ JS-Offer: ' + generate_random_number(4);
+        affiliate_offer.bml_id = utils().generate_random_number(10);
+        affiliate_offer.shopping_title = '$$ JS-Offer: ' + utils().generate_random_number(4);
         affiliate_offer.expiration_date = Date.tomorrow();
         affiliate_offer.shopping_distribution.start_date = new Date();
         affiliate_offer.shopping_distribution.end_date = Date.tomorrow();
@@ -28,18 +29,12 @@ module.exports = function offer(nemo) {
 
     };
 
-    var generate_random_number = function(length) {
-        var begin = 2;
-        return Math.random().toString().slice(begin,length + begin);
-    };
 
     var create_affiliate_offer = function() {
-
         return setup_affiliate_offer()
             .then(submit_affiliate_offer);
 
         function submit_affiliate_offer(affiliate_offer) {
-
             return client(nemo)
                 .submit_affiliate_offer(affiliate_offer);
         }
